backend/app/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Template for backend with FastAPI and SQLModel"
authors = ["ikrzywda <krzywdaigor890@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.98.0"
pydantic = {version = "^1.10.9", extras = ["email"]}
alembic = "^1.11.1"
sqlmodel = "^0.0.8"
psycopg2-binary = "^2.9.6"
requests = "^2.31.0"
SQLAlchemy = "^1.4.41"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
asyncpg = "^0.28.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
python-multipart = "^0.0.6"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.4.1"
black = "^23.3.0"
ruff = "^0.0.275"
sqlalchemy-stubs = "^0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["pydantic.mypy", "sqlmypy"]
ignore_missing_imports = true
disallow_untyped_defs = true
